training/requirements.txt
tensorflow
keras2onnx
onnxruntime
pyyaml
